% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindif.R
\name{bindif}
\alias{bindif}
\title{Determine the difference between a binomial sample and dbinom
Modified version of mybin from the lab5 help files}
\usage{
bindif(iter = 10000, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of data points in the sample}

\item{n}{number of success}

\item{p}{probability (0 < p < 1)}
}
\value{
a dataframe table showing the differences
}
\description{
Determine the difference between a binomial sample and dbinom
Modified version of mybin from the lab5 help files
}
