% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyshev.R
\name{chebyshev}
\alias{chebyshev}
\title{Determine the number of observations within a certain standard deviation for lab 2}
\usage{
chebyshev(sd, vector)
}
\arguments{
\item{sd}{number of standard deviations away from the mean}

\item{vector}{vector of data}
}
\value{
the percentage of data that falls within a specified number of standard deviations from the data
}
\description{
More detailed description
}
\examples{
chebyshev(2, c(1,2,3,4,5,6,7,8,9,11,13,16,18,21))
chebyshev(1, c(1,2,3,11,13,16,18,21,34,41,56))
}
