% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fire.r
\docType{data}
\name{fire}
\alias{fire}
\title{Fire damage}
\format{
A data frame with 15 rows and 2 variables:
\describe{
\item{DISTANCE}{Distance of the house from the fire station in miles}
\item{DAMAGE}{Damage in thousands of dollars}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set collected concerning the cst of home damage from fire
}
\keyword{datasets}
