% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCompare.R
\name{fitCompare}
\alias{fitCompare}
\title{Determine which of two models has the better fit}
\usage{
fitCompare(m1, m2)
}
\arguments{
\item{m1}{model 1}

\item{m2}{model 2}
}
\value{
returns a vector with value[1] = TSS, value[2] = MSS, value[3] = RSS
}
\description{
More detailed description
}
