% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Create a 95\% confidence interval for mu given a sample normally distributed x}
\usage{
myci(x)
}
\arguments{
\item{x}{a vector sample from a population}
}
\value{
returns a list containing the confidence interval
}
\description{
More detailed description
}
