% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Print out a curve like in lab 6.
display the curve, shaded area between the curve and x axis from to x=a, and
calculate the area ??????(probability, P(X<=a)) which is released to the command line}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{x value for the graph such that the graph will display values -inf < x < a}
}
\value{
a graph
}
\description{
Print out a curve like in lab 6.
display the curve, shaded area between the curve and x axis from to x=a, and
calculate the area ??????(probability, P(X<=a)) which is released to the command line
}
