% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nathanboot2.R
\name{nathanboot2}
\alias{nathanboot2}
\title{Create a Bootstrap Confidence Interval for a given sample}
\usage{
nathanboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iteration to sample from}

\item{x}{the sample from which data will be sampled}

\item{fun}{the name of the function to be applied to the data, for example, mean, sd, median, etc...}

\item{alpha}{the alpha value used to construct confidence intervals of size (1-alpha)*100}

\item{cx}{numerical character expansion factor}

\item{...}{additional arguments for formatting the histogram}
}
\value{
returns a list containing the confidence interval, the function used, and the sample
}
\description{
More detailed description
}
