% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{nathancltp}
\alias{nathancltp}
\title{Create various plots for CLT of a poisson distribution.}
\usage{
nathancltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{the number of samples to be created for each iteration}

\item{iter}{the number of iterations to be created}

\item{lambda}{the lambda value for the poisson distribution}

\item{...}{additional arguments}
}
\value{
This function returns three graphs: a histogram of the sample mean, a theoretical normal curve, and a probability function bar plot, all with rainbow colors
}
\description{
Create various plots for CLT of a poisson distribution.
}
\examples{
nathancltp(50, 1000, 6)
nathancltp(10, 6000)
nathancltp(20, 100, 2)
}
