% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{nathanmlnorm}
\alias{nathanmlnorm}
\title{Determines the mu estimate and sigma estimate for a normal distribution usinig the maximum likelihood method.}
\usage{
nathanmlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{a vector of samples from a normal distribution}

\item{mu}{a sequence of possible mu values from the minimum value in x to the maximum value in x}

\item{sig}{a sequence of possible sigma values within which the real sigma should exist}

\item{...}{other graphical parameters}
}
\value{
a list of objects, including the sample, the coordiates of the estimated mu and sigma, the mu estimate, and the sigma estimate
}
\description{
Determines the mu estimate and sigma estimate for a normal distribution usinig the maximum likelihood method.
}
