% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nathanpvalue.R
\name{nathanpvalue}
\alias{nathanpvalue}
\title{Create various plots for CLT of a poisson distribution.}
\usage{
nathanpvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{the calculated t statistic}

\item{xmax}{the highest possible t statistic}

\item{n}{the number of samples to create}

\item{alpha}{alpha value for a 1-alpha percent confidence interval}
}
\value{
This function returns a graph of the t statistic with the sample mean and the area described, and a list containing the bounds on the rejection region as well as the p value
}
\description{
Create various plots for CLT of a poisson distribution.
}
