% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmr.R
\name{tmr}
\alias{tmr}
\title{Determine TSS, MSS, and RSS for lab 3}
\usage{
tmr(y, x, dataArg)
}
\arguments{
\item{y}{the y variable of the model}

\item{x}{the x variable of the model}

\item{dataArg}{the dataframe for the model}
}
\value{
returns a vector with value[1] = TSS, value[2] = MSS, value[3] = RSS
}
\description{
More detailed description
}
