% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violPlot.R
\name{violPlot}
\alias{violPlot}
\title{Create a violin ggplot for lab 1}
\usage{
violPlot(df, x, y, z)
}
\arguments{
\item{df}{dataframe}

\item{x}{column of dataframe used as the horizontal axis}

\item{y}{column of dataframe used as the vertical axis}

\item{z}{column of dataframe used as the fill for the violin plot}
}
\value{
a ggplot object
}
\description{
Create a violin ggplot for lab 1
}
