% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyshev.R
\name{chebyshev}
\alias{chebyshev}
\title{Determine the number of observations within a certain standard deviation for lab 2}
\usage{
chebyshev(sd, vector)
}
\arguments{
\item{sd}{number of standard deviations away from the mean}

\item{vector}{vector of data}
}
\value{
a ggplot object
}
\description{
More detailed description
}
\details{
More detailed description
}
